Created service for products and warehouses

+++ order-stock-microservices/product-inventory-service/src/main/java/org/productinventoryservice/dto/product/DeleteProductDTO.java
@@ -1,4 +1,4 @@
 package org.productinventoryservice.dto.product;
 
-public record DeleteProductDTO(Integer id) {
+public record DeleteProductDTO(Long id) {
 }
